So much fixes, so wow...
Use Device Mode enums everywhere, versus integers. Because that's why we made them.

Ensure auto-disable/enable works properly and persists across reboots.

Bump DreamScreenNet package version, which fixes extraneous logging to syslog.

+++ src/Glimmr/Controllers/GlimmrController.cs
@@ -429,7 +429,7 @@ namespace Glimmr.Controllers {
 		[HttpPost("mode")]
 		public async Task<ActionResult<DeviceMode>> DevMode([FromBody] DeviceMode mode) {
 			Log.Debug("Mode set to: " + mode);
-			await _controlService.SetMode((int)mode);
+			await _controlService.SetMode(mode);
 			return new ActionResult<DeviceMode>(mode);
 		}
